refactor: change variable type to long

+++ src/main/java/com/craftelix/dto/MatchFilterResponseDto.java
@@ -13,7 +13,7 @@ public class MatchFilterResponseDto {
 
     private int page;
 
-    private long pageCount;
+    private int pageCount;
 
     private String playerName;
